+++ assets/javascripts/kitten/components/lists/list/components/button-item.js
@@ -26,6 +26,12 @@ const Item = styled.div`
     outline: ${COLORS.primary4} solid ${pxToRem(2)};
     outline-offset: ${pxToRem(2)};
   }
+  &:focus:not(:focus-visible) {
+    outline-color: transparent;
+  }
+  &:focus-visible {
+    outline-color: ${COLORS.primary4};
+  }
 
   &.k-List__button--withTopBorder {
     border-top: ${pxToRem(2)} solid ${COLORS.line1};