+++ assets/javascripts/kitten/components/steppers/stepper/index.js
@@ -77,6 +77,12 @@ const StepperText = styled(Text)`
     outline: ${COLORS.primary3} solid ${pxToRem(2)};
     outline-offset: ${pxToRem(2)};
   }
+  &:focus:not(:focus-visible) {
+    outline-color: transparent;
+  }
+  &:focus-visible {
+    outline-color: ${COLORS.primary4};
+  }
 `
 
 export const StepperItem = ({ children, state, ...props }) => {