+++ assets/javascripts/kitten/components/uploaders/basic-uploader/index.js
@@ -39,6 +39,12 @@ const StyledBasicUploader = styled.div`
       fill: ${COLORS.background1};
     }
   }
+  input[type='file']:focus:not(:focus-visible) + label {
+    outline-color: transparent;
+  }
+  input[type='file']:focus-visible + label {
+    outline-color: ${COLORS.primary4};
+  }
 
   &:not(.k-BasicUploader--loading) {
     input[type='file']:disabled + label {
@@ -76,6 +82,12 @@ const StyledBasicUploader = styled.div`
     &:focus {
       outline: ${COLORS.primary4} solid ${pxToRem(2)};
       outline-offset: ${pxToRem(2)};
+    }
+    &:focus:not(:focus-visible) {
+      outline-color: transparent;
+    }
+    &:focus-visible {
+      outline-color: ${COLORS.primary4};
     }
   }
 `